Updated database dump. Fixed bugs in Ship endpoint.

+++ server/src/main/java/com/server/server/repository/ShipRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShipRepository extends JpaRepository<Ship, Long> {
+public interface ShipRepository extends JpaRepository<Ship, Integer> {
 
 }
